Fixed incorrect casting to int for profile ids. Fixed incorrect URL for reordering.
git-svn-id: http://code.elgg.org/elgg/trunk@5656 36083f99-b078-4883-b0ff-0f9b5a30f544

+++ mod/profile/actions/deletedefaultprofileitem.php
@@ -13,7 +13,7 @@ global $CONFIG;
 
 admin_gatekeeper();
 
-$id = (int)get_input('id');
+$id = get_input('id');
 
 $fieldlist = get_plugin_setting('user_defined_fields', 'profile');
 if (!$fieldlist) {
